Basic Makefile added
This is a simple Makefile to be used to build libfms and the ocean-only
and ocean-ice MOM6 executables.  It currently does not explicitly
reference the actual files, and is mostly just a direct translation of
the instructions from the "Getting Started" wiki.

Unlikely this will ever be merged, the purpose of this commit is mostly
to allow me to keep track of the file (and possibly preserve it).