Partial update to support mainline FMS
Rules for building the mainline FMS library have been added to the
top Makefile.  New preprocessor types have been added to the FMS
makefile.

What has *not* been added is the required changed to makedep to support
this build.  This needs to go through MOM6 unfortunately.

This commit is not ready; expect it to be squashed/rebased.