Simpler Makefile; more modular ocean_only Makefile
The top-level Makefile is now a simpler thing that does not support
BUILD or symmetric/asymmetric builds.  It now simply goes inside of the
respective model directories and builds the contents.  The more
elaborate configurations are now to be handled at the platform driver
(e.g. Gaea-stats).

The ocean_only makefile was modified to be more responsive to file
changes.  If a MOM source is changed, it rebuilds.  If libFMS changes,
then it wipes and rebuilds MOM6.  For now, it also re-runs configure.

There was a bit of splitting between configure and config.status in the
ocean_only makefile, although I doubt it really added much value.