get_nprocs.py: Replacement CPU parser
The get_nprocs.sh script has been replaced with get_nprocs.py, a Python
script which parses both MOM_parameter_doc.layout and input.nml to
determine the total number of CPUs for a given experiment.

The "maxiumum CPU limit" is now determined by the psutils Python module.
It will only be used if the module is available.  There is no true
platform-independent or even POSIX way to determine the total CPU count,
but psutil provides a way to do this on a variety of platforms.  If it's
available.