parse_mom_input.py: MOM parameter doc file parser
This script parses a MOM parameter file and can be used as either a
Python module or a command line tool.

If used as a module, it returns a dict of all Python parameters.  AFAIK
nearly everything is supported, including blocks and repeat tokens (e.g.
x = 5*1).  Only A%B%C = 1 is not yet supported.

As a command line tool, you can provide the file and a parameter and it
will return the value.  Probably won't work for blocks.  I don't think
there are any plans to use it as a CLI tool.

The f90nml Python module was added as a submodule to support this and
future support scripts.