gen_input_manifest.py: Experiment input file list
This script is a command line too which attempts to generate the list of
input files for a given experiment.

Information is taken from both input.nml and MOM parameter files to
reconstruct the list of input files, along with any expected FMS files.
Any information in INPUT (or perhaps the param's INPUTDIR) is assumed to
be a symbolic link and is moved over by the associated Makefile.

This script is far from perfect, and it is perhaps impossible to ever
create such a script, but it should be sufficient for this repo.