Update Gitlab config to new Makefile testing
The Gitlab configuration file is updated to use the new Makefile-based
testing.  Multiple tasks which were previously defined in
tools/MRS/Makefile are moved into .gitlab-ci.yml.

While still not a platform-independent setup, there is at least now a
logical separation of the environment from the building and running of
the models.