Force memory for restarts and sis2 runs
The default memory layout for restarts is changed to symmetric.  Certain
configurations are not able to generate reproducible runs in
nonsymmetric mode, so this will help ensure that they use symmetric.

SIS2 nonsymmetric regression test output is now explicitly set as
nonsymmetric.  This ensures nonsymmetric layout if the MEMORY macro is
changed.